+++ test/int/sql/test_datetime.py
@@ -1,0 +1,9 @@
+from framework.sqltester import (
+    ClusterTwoInstances,
+    sql_test_file,
+)
+
+
+@sql_test_file("test_datetime.sql")
+class TestDatetime(ClusterTwoInstances):
+    pass